Showing issues with either/both promotion and motion-kit also motion layout when used with Promotion
Uncomment the desired test in app_delegate.
InquiryList uses PM::TableViewCell and cells are not the proper height. sliding left for delete shows the proper height of the delete button. Also rotating to landscape the background is only the width of the portrait.  Continuing to rotate around leaves in landscape when in reversed portrait.

MyAwsomeScreen uses a UITableViewCell directly including PM::TableViewCell and has the same problems

The MotionLayoutScreen gives a 'NoMethodError', reason: 'table.rb:120:in `create_table_cell:': undefined method `autoresizingMask=' for #<Motion::Layout:0x11769200 ...>

error on compiling.